Changed the timer resolution from 1 second to 0.1 sec.

+++ app/src/main/java/com/billweberx/pt_timer/data/TimerScreenState.kt
@@ -3,7 +3,7 @@ package com.billweberx.pt_timer.data
 // Data class for the main timer screen's reactive state, moved here to be globally accessible
 data class TimerScreenState(
     val status: String = "Ready",
-    val remainingTime: Int = 0,
+    val remainingTime: Long = 0,
     val progressDisplay: String = "",
     val currentSet: Int = 0,
     val currentRep: Int = 0,
